requirements.txt
altair==5.5.0
altex==0.2.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.39.12
botocore==1.39.12
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.0
cryptography==45.0.5
cycler==0.12.1
deprecation==2.1.0
entrypoints==0.4
exceptiongroup==1.3.0
Faker==37.4.2
favicon==0.7.0
filelock==3.18.0
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.45
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
htbuilder==0.9.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
lxml==6.0.0
Markdown==3.8.2
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.9.4
narwhals==1.48.0
numpy==2.0.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
postgrest==1.1.1
prometheus_client==0.22.1
protobuf==6.31.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pymdown-extensions==10.16
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
realtime==2.6.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
s3transfer==0.13.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.16.0
snowflake-snowpark-python==1.34.0
sortedcontainers==2.4.0
soupsieve==2.7
st-annotated-text==4.0.2
st-theme==1.2.3
storage3==0.12.0
streamlit==1.47.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.5
streamlit-image-coordinates==0.2.2
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.20
validators==0.35.0
watchdog==6.0.0
websockets==15.0.1
zipp==3.23.0
